requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=4.0.0
datasets>=2.0.0
 crewai>=0.11.0  
 ragas>=0.4.0    
unstructured[pdf]>=0.10.0